Move: 個人のプロフィールのモデルを別ファイルに分けた
